backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
openai==1.12.0
dj-database-url==2.1.0
gunicorn==21.2.0
pytest==7.4.4
pytest-django==4.8.0
